secret_manager/requirements.txt
fastapi
uvicorn[standard]
hvac
requests
python-jose
python-keycloak
pydantic
